+++ src/scooze/routers/decks.py
@@ -1,9 +1,10 @@
 from typing import Any
 
-import scooze.database.deck as db
-from fastapi import APIRouter
+from beanie import PydanticObjectId
+from fastapi import APIRouter, HTTPException
 from fastapi.responses import JSONResponse
-from scooze.models.deck import DeckModelIn, DeckModelOut
+from scooze.models.deck import DeckModel, DeckModelData
+from scooze.utils import to_lower_camel
 
 router = APIRouter(
     prefix="/decks",
@@ -13,83 +14,115 @@ router = APIRouter(
 
 
 @router.get("/", summary="Get decks at random")
-async def decks_root(limit: int = 3) -> DeckModelOut:
+async def decks_root(limit: int = 3) -> list[DeckModel]:
     """
-    Get random decks up to the given limit.
+    Get random decks from the database.
 
     Args:
-        limit: the maximum number of decks to get
+        limit: The maximum number of decks to get.
+
+    Returns:
+        Random decks from the database.
+
+    Raises:
+        HTTPException: 404 - No decks found in the database.
     """
 
-    decks = await db.get_decks_random(limit=limit)
-    if decks is not None:
-        return JSONResponse([deck.model_dump(mode="json") for deck in decks], status_code=200)
-    else:
-        return JSONResponse({"message": "No decks found in the database."}, status_code=404)
+    decks = await DeckModel.aggregate([{"$sample": {"size": limit}}], projection_model=DeckModel).to_list()
+
+    if decks is None or not decks:
+        raise HTTPException(status_code=404, detail="No decks found in the database.")
+
+    return decks
 
 
 # Create
 
 
 @router.post("/add", summary="Create new decks")
-async def add_decks(decks: list[DeckModelIn]) -> DeckModelOut:
+async def add_decks(decks: list[DeckModelData]) -> JSONResponse:
     """
-    Add decks to the database.
+    Add a list of decks to the database.
 
     Args:
-        decks: The decks to add.
+        decks: A list of dicts conforming to DeckModelData's schema.
 
     Returns:
-        The scooze IDs of the created decks.
+        A message stating how many decks were created.
 
     Raises:
         HTTPException: 400 - Create failed, passes along the error message.
     """
 
-    inserted_ids = await db.add_decks(decks=decks)
-    if inserted_ids is not None:
-        return JSONResponse({"message": f"Created {len(inserted_ids)} deck(s)."}, status_code=200)
-    else:
-        return JSONResponse({"message": "Failed to create any new decks."}, status_code=400)
+    try:
+        decks_to_insert = [DeckModel.model_validate(deck.model_dump()) for deck in decks]
+        insert_result = await DeckModel.insert_many(decks_to_insert)
+        return JSONResponse(f"Created {len(insert_result.inserted_ids)} deck(s).")
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f"Failed to create new decks.")
 
 
 @router.post("/by", summary="Get decks by property")
 async def get_decks_by(
-    property_name: str, values: list[Any], paginated: bool = False, page: int = 1, page_size: int = 10
-) -> DeckModelOut:
+    property_name: str,
+    values: list[Any],
+    paginated: bool = False,
+    page: int = 1,
+    page_size: int = 10,
+) -> list[DeckModel]:
     """
     Get decks where the given property matches any of the given values.
 
     Args:
-        property_name: the property to check against
-        values: matching values of the given property
-        paginated: return paginated results if True, return all matches if
-            False
-        page: return matches from the given page
-        page_size: the number of results per page
+        property_name: The property to check against.
+        values: Matching values of the given property.
+        paginated: Return paginated results if True, or all matches if False.
+        page: The page to return matches from.
+        page_size: The number of results per page.
+
+    Returns:
+        A list of decks matching the search criteria.
+
+    Raises:
+        HTTPException: 404 - Decks weren't found.
     """
 
-    decks = await db.get_decks_by_property(
-        property_name=property_name, values=values, paginated=paginated, page=page, page_size=page_size
-    )
-    if decks is not None:
-        return JSONResponse([deck.model_dump(mode="json") for deck in decks], status_code=200)
-    else:
-        return JSONResponse({"message": "Decks not found."}, status_code=404)
+    match property_name:
+        case "_id" | "id":
+            prop_name = "_id"
+            vals = [PydanticObjectId(v) for v in values]  # Normalize Mongo IDs
+        case _:
+            prop_name = to_lower_camel(property_name)
+            vals = values
+
+    skip = (page - 1) * page_size if paginated else 0
+    limit = page_size if paginated else None
+    decks = await DeckModel.find({"$or": [{prop_name: v} for v in vals]}, skip=skip, limit=limit).to_list()
+
+    if len(decks) == 0:
+        raise HTTPException(status_code=404, detail="Decks not found.")
+
+    return decks
 
 
 # Delete
 
 
 @router.delete("/delete/all", summary="Delete all decks")
-async def delete_decks_all() -> DeckModelOut:
+async def delete_decks_all() -> JSONResponse:
     """
-    Delete all decks.
+    Deletes all decks in the database.
+
+    Returns:
+        A message that the decks were deleted.
+
+    Raises:
+        HTTPException: 400 - Decks weren't deleted.
     """
 
-    deleted_count = await db.delete_decks_all()
+    delete_result = await DeckModel.delete_all()
 
-    if deleted_count is not None:
-        return JSONResponse({"message": f"Deleted {deleted_count} deck(s)."}, status_code=200)
-    else:
-        return JSONResponse({"message": "No decks deleted."}, status_code=404)
+    if delete_result is None:
+        raise HTTPException(status_code=400, detail="Decks weren't deleted.")
+
+    return JSONResponse(f"Deleted {delete_result.deleted_count} deck(s).")